Update README.md
The purpose of this dataset is to facilitate the development and evaluation of machine learning models for car price prediction. The dataset can be used for tasks such as regression analysis, feature engineering, and model training.

Please note that the dataset is collected from a specific source and represents a snapshot of used car listings at a particular point in time. The dataset may contain missing values and some discrepancies. It is recommended to perform data preprocessing and exploratory analysis before using it for research or predictive modeling.

We hope that this dataset serves as a valuable resource for the data science community and contributes to the advancement of automotive analytics and pricing models.